[Redux] Rename queriesSlice to transactionsSlice
