+++ packages/desktop-client/src/components/mobile/transactions/TransactionListItem.tsx
@@ -95,7 +95,9 @@ export function TransactionListItem({
   const transferAccount = useAccount(payee?.transfer_acct || '');
   const isPreview = isPreviewId(transaction?.id || '');
 
-  const newTransactions = useSelector(state => state.queries.newTransactions);
+  const newTransactions = useSelector(
+    state => state.transactions.newTransactions,
+  );
 
   const { longPressProps } = useLongPress({
     accessibilityDescription: 'Long press to select multiple transactions',
