+++ packages/desktop-client/src/components/modals/ImportTransactionsModal/ImportTransactionsModal.jsx
@@ -41,11 +41,11 @@ import { useCategories } from '@desktop-client/hooks/useCategories';
 import { useDateFormat } from '@desktop-client/hooks/useDateFormat';
 import { useSyncedPrefs } from '@desktop-client/hooks/useSyncedPrefs';
 import { reloadPayees } from '@desktop-client/payees/payeesSlice';
+import { useDispatch } from '@desktop-client/redux';
 import {
   importPreviewTransactions,
   importTransactions,
-} from '@desktop-client/queries/queriesSlice';
-import { useDispatch } from '@desktop-client/redux';
+} from '@desktop-client/accounts/accountssSlice';
 
 function getFileType(filepath) {
   const m = filepath.match(/\.([^.]*)$/);
